Added initial proccesing and ability to get chunk count on a job through jobstore.
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@60752 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ engine-prototype/src/main/java/dk/dbc/dataio/engine/JobStore.java
@@ -7,5 +7,5 @@ interface JobStore {
     
     void addChunk(Job job, Chunk chunk) throws JobStoreException;
 
-    int getNumberOfChunksInJob(Job job) throws JobStoreException;
+    long getNumberOfChunksInJob(Job job) throws JobStoreException;
 }
